feat: added static ticket type, also totp utilities and both static and dynamic ticket frontend

+++ src/Domain/EventManagement/Actions/DeleteEventProduct.php
@@ -13,6 +13,11 @@ class DeleteEventProduct
     {
         try {
             $product = Product::findOrFail($productId);
+
+            if ($product->tickets()->exists()) {
+                return false;
+            }
+
             $product->delete();
             return true;
         } catch (\Exception $e) {
@@ -20,10 +25,13 @@ class DeleteEventProduct
         }
     }
 
-    // TODO: if tickets exists, cant delete product
     public function asController($eventId, $productId)
     {
-        $this->handle($eventId, $productId);
+        $result = $this->handle($eventId, $productId);
+
+        if (!$result) {
+            return back()->with('message', flash_error('Product cannot be deleted', 'Product cannot be deleted because it has tickets'));
+        }
 
         return back()->with('message', flash_success('Product deleted successfully', 'Product deleted successfully'));
     }
